requirements.txt
flask
requests
newspaper3k
transformers
torch